Add flush directory method

+++ lib/StorageArea.php
@@ -79,11 +79,16 @@ class StorageArea
     /**
      * Remove all files and directories in the current storage
      *
+     * @param  string $path Remove all files and directories from given path
      * @return void
      */
-    public function flush() {
-        foreach ($this->nodes as $node) {
-            $node->delete();
+    public function flush($path = null) {
+        if ($path) {
+            StorageDirectory::flush($path);
+        } else {
+            foreach ($this->nodes as $node) {
+                $node->delete();
+            }
         }
     }
 }
